+++ src/main/scala/za/co/absa/abris/examples/ConfluentKafkaAvroReader.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 ABSA Group Limited
+ * Copyright 2019 ABSA Group Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,18 +18,15 @@ package za.co.absa.abris.examples
 
 import java.util.Properties
 
-import org.apache.spark.sql.streaming.DataStreamReader
-import org.apache.spark.sql.{Dataset, Row, SparkSession}
-import za.co.absa.abris.avro.schemas.policy.SchemaRetentionPolicies.{RETAIN_ORIGINAL_SCHEMA, RETAIN_SELECTED_COLUMN_ONLY}
+import org.apache.spark.sql.functions.col
+import org.apache.spark.sql.{DataFrame, Dataset, Row}
 import za.co.absa.abris.examples.utils.ExamplesUtils._
 
 object ConfluentKafkaAvroReader {
 
   private val PARAM_JOB_NAME = "job.name"
   private val PARAM_JOB_MASTER = "job.master"
-  private val PARAM_KEY_AVRO_SCHEMA = "key.avro.schema"
   private val PARAM_PAYLOAD_AVRO_SCHEMA = "payload.avro.schema"
-  private val PARAM_TASK_FILTER = "task.filter"
   private val PARAM_LOG_LEVEL = "log.level"
   private val PARAM_OPTION_SUBSCRIBE = "option.subscribe"
 
@@ -47,9 +44,10 @@ object ConfluentKafkaAvroReader {
     val stream = spark
       .readStream
       .format("kafka")
-      .addOptions(properties) // 1. this method will add the properties starting with "option."; 2. security options can be set in the properties file
+      .addOptions(properties) // 1. this method will add the properties starting with "option."
+                              // 2. security options can be set in the properties file
 
-    val deserialized = configureExample(stream, properties)
+    val deserialized = configureExample(stream.load(), properties)
 
     // YOUR OPERATIONS CAN GO HERE
 
@@ -58,17 +56,22 @@ object ConfluentKafkaAvroReader {
     deserialized
       .writeStream
       .format("console")
+      .option("truncate", "false")
       .start()
       .awaitTermination()
   }
 
-  private def configureExample(stream: DataStreamReader,props: Properties): Dataset[Row] = {
-    import za.co.absa.abris.avro.AvroSerDe._
-    if (props.getProperty(PARAM_EXAMPLE_SHOULD_USE_SCHEMA_REGISTRY).toBoolean) {
-      stream.fromConfluentAvro("value", None, Some(props.getSchemaRegistryConfigurations(PARAM_OPTION_SUBSCRIBE)))(RETAIN_ORIGINAL_SCHEMA)
-    }
-    else {
-      stream.fromConfluentAvro("value", Some(props.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA)), None)(RETAIN_SELECTED_COLUMN_ONLY)
+  private def configureExample(dataFrame: DataFrame, properties: Properties): Dataset[Row] = {
+
+    import za.co.absa.abris.avro.functions.from_confluent_avro
+
+    if (properties.getProperty(PARAM_EXAMPLE_SHOULD_USE_SCHEMA_REGISTRY).toBoolean) {
+      val schemaRegistryConfig = properties.getSchemaRegistryConfigurations(PARAM_OPTION_SUBSCRIBE)
+      dataFrame.select(from_confluent_avro(col("value"), schemaRegistryConfig) as 'data).select("data.*")
+    } else {
+      val source = scala.io.Source.fromFile(properties.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA))
+      val schemaString = try source.mkString finally source.close()
+      dataFrame.select(from_confluent_avro(col("value"), schemaString) as 'data).select("data.*")
     }
   }
 }
