+++ src/main/scala/za/co/absa/abris/examples/KafkaAvroReader.scala
@@ -18,20 +18,15 @@ package za.co.absa.abris.examples
 
 import java.util.Properties
 
-import org.apache.spark.sql.streaming.DataStreamReader
-import org.apache.spark.sql.{Dataset, Row, SparkSession}
-import za.co.absa.abris.avro.schemas.policy.SchemaRetentionPolicies.RETAIN_SELECTED_COLUMN_ONLY
+import org.apache.spark.sql.DataFrame
+import org.apache.spark.sql.functions.col
 import za.co.absa.abris.examples.utils.ExamplesUtils._
-
-
 
 object KafkaAvroReader {
 
   private val PARAM_JOB_NAME = "job.name"
   private val PARAM_JOB_MASTER = "job.master"
-  private val PARAM_KEY_AVRO_SCHEMA = "key.avro.schema"
   private val PARAM_PAYLOAD_AVRO_SCHEMA = "payload.avro.schema"
-  private val PARAM_TASK_FILTER = "task.filter"
   private val PARAM_LOG_LEVEL = "log.level"
   private val PARAM_OPTION_SUBSCRIBE = "option.subscribe"
 
@@ -50,31 +45,33 @@ object KafkaAvroReader {
     val stream = spark
       .readStream
       .format("kafka")
-      .addOptions(properties) // 1. this method will add the properties starting with "option."; 2. security options can be set in the properties file
+      .addOptions(properties) // 1. this method will add the properties starting with "option."
+                              // 2. security options can be set in the properties file
 
-    val deserialized = configureExample(stream, properties)
+    val deserialized = configureExample(stream.load(), properties)
 
     // YOUR OPERATIONS CAN GO HERE
 
     deserialized.printSchema()
-
     deserialized
-    .writeStream
+      .writeStream
       .format("console")
+      .option("truncate", "false")
       .start()
       .awaitTermination()
   }
 
-
-
-  private def configureExample(stream: DataStreamReader,props: Properties): Dataset[Row] = {
-    import za.co.absa.abris.avro.AvroSerDe._
+  private def configureExample(dataFrame: DataFrame, props: Properties): DataFrame = {
+    import za.co.absa.abris.avro.functions.from_avro
 
     if (props.getProperty(PARAM_EXAMPLE_SHOULD_USE_SCHEMA_REGISTRY).toBoolean) {
-      stream.fromAvro("value", props.getSchemaRegistryConfigurations(PARAM_OPTION_SUBSCRIBE))(RETAIN_SELECTED_COLUMN_ONLY)
+      val schemaRegistryConfig = props.getSchemaRegistryConfigurations(PARAM_OPTION_SUBSCRIBE)
+      dataFrame.select(from_avro(col("value"), schemaRegistryConfig) as 'data).select("data.*")
     }
     else {
-      stream.fromAvro("value", props.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA))(RETAIN_SELECTED_COLUMN_ONLY)
+      val source = scala.io.Source.fromFile(props.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA))
+      val schemaString = try source.mkString finally source.close()
+      dataFrame.select(from_avro(col("value"), schemaString) as 'data).select("data.*")
     }
   }
 }
