+++ src/main/scala/za/co/absa/abris/examples/deprecated/using_keys/ConfluentKafkaAvroReaderWithKey.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 ABSA Group Limited
+ * Copyright 2018 ABSA Group Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,24 @@
  * limitations under the License.
  */
 
-package za.co.absa.abris.examples.sql
+package za.co.absa.abris.examples.deprecated.using_keys
 
 import java.util.Properties
 
-import org.apache.spark.sql.functions.col
-import org.apache.spark.sql.{DataFrame, Dataset, Row}
+import org.apache.spark.sql.streaming.DataStreamReader
+import org.apache.spark.sql.{Dataset, Row, SparkSession}
+import za.co.absa.abris.avro.schemas.policy.SchemaRetentionPolicies.{RETAIN_ORIGINAL_SCHEMA, RETAIN_SELECTED_COLUMN_ONLY}
 import za.co.absa.abris.examples.utils.ExamplesUtils._
 
-object NewConfluentKafkaAvroReader {
+import scala.collection.JavaConversions._
+
+object ConfluentKafkaAvroReaderWithKey {
 
   private val PARAM_JOB_NAME = "job.name"
   private val PARAM_JOB_MASTER = "job.master"
+  private val PARAM_KEY_AVRO_SCHEMA = "key.avro.schema"
   private val PARAM_PAYLOAD_AVRO_SCHEMA = "payload.avro.schema"
+  private val PARAM_TASK_FILTER = "task.filter"
   private val PARAM_LOG_LEVEL = "log.level"
   private val PARAM_OPTION_SUBSCRIBE = "option.subscribe"
 
@@ -44,12 +49,11 @@ object NewConfluentKafkaAvroReader {
     val stream = spark
       .readStream
       .format("kafka")
-      .addOptions(properties) // 1. this method will add the properties starting with "option."
-                              // 2. security options can be set in the properties file
+      .addOptions(properties) // 1. this method will add the properties starting with "option."; 2. security options can be set in the properties file
 
-    val deserialized = configureExample(stream.load(), properties)
+    val deserialized = configureExample(stream, properties)
 
-    // YOUR OPERATIONS CAN GO HERE
+    // YOUR CODE CAN COME HERE
 
     deserialized.printSchema()
 
@@ -61,17 +65,15 @@ object NewConfluentKafkaAvroReader {
       .awaitTermination()
   }
 
-  private def configureExample(dataFrame: DataFrame, properties: Properties): Dataset[Row] = {
+  private def configureExample(stream: DataStreamReader,props: Properties): Dataset[Row] = {
 
-    import za.co.absa.abris.avro.functions.from_confluent_avro
+    import za.co.absa.abris.avro.AvroSerDeWithKeyColumn._
 
-    if (properties.getProperty(PARAM_EXAMPLE_SHOULD_USE_SCHEMA_REGISTRY).toBoolean) {
-      val schemaRegistryConfig = properties.getSchemaRegistryConfigurations(PARAM_OPTION_SUBSCRIBE)
-      dataFrame.select(from_confluent_avro(col("value"), schemaRegistryConfig) as 'data).select("data.*")
-    } else {
-      val source = scala.io.Source.fromFile(properties.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA))
-      val schemaString = try source.mkString finally source.close()
-      dataFrame.select(from_confluent_avro(col("value"), schemaString) as 'data).select("data.*")
+    if (props.getProperty(PARAM_EXAMPLE_SHOULD_USE_SCHEMA_REGISTRY).toBoolean) {
+      stream.fromConfluentAvro(props.getSchemaRegistryConfigurations(PARAM_OPTION_SUBSCRIBE))(RETAIN_ORIGINAL_SCHEMA)
+    }
+    else {
+      stream.fromConfluentAvro(props.getProperty(PARAM_KEY_AVRO_SCHEMA), props.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA), props.getSchemaRegistryConfigurations(PARAM_OPTION_SUBSCRIBE))(RETAIN_SELECTED_COLUMN_ONLY)
     }
   }
 }
