+++ src/main/scala/za/co/absa/abris/examples/deprecated/using_keys/ConfluentKafkaAvroWriterWithKey.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package za.co.absa.abris.examples.using_keys
+package za.co.absa.abris.examples.deprecated.using_keys
 
 import java.io.FileInputStream
 import java.util.Properties
@@ -25,11 +25,12 @@ import org.apache.spark.sql.types.{StructField, StructType}
 import org.apache.spark.sql.{Dataset, Encoder, Row, SparkSession}
 import za.co.absa.abris.avro.format.SparkAvroConversions
 import za.co.absa.abris.avro.parsing.utils.AvroSchemaUtils
+import za.co.absa.abris.avro.read.confluent.SchemaManager
 import za.co.absa.abris.examples.data.generation.ComplexRecordsGenerator
 import za.co.absa.abris.examples.utils.ExamplesUtils._
 import scala.collection.JavaConversions._
 
-object KafkaAvroWriterWithKey {
+object ConfluentKafkaAvroWriterWithKey {
 
   private val PARAM_JOB_NAME = "job.name"
   private val PARAM_JOB_MASTER = "job.master"
@@ -42,6 +43,7 @@ object KafkaAvroWriterWithKey {
   private val PARAM_TEST_DATA_ENTRIES = "test.data.entries"
   private val PARAM_EXECUTION_REPEAT = "execution.repeat"
   private val PARAM_NUM_PARTITIONS = "num.partitions"
+  private val PARAM_TOPIC = "option.topic"
 
   def main(args: Array[String]): Unit = {
 
@@ -53,7 +55,6 @@ object KafkaAvroWriterWithKey {
     val spark = getSparkSession(properties, PARAM_JOB_NAME, PARAM_JOB_MASTER, PARAM_LOG_LEVEL)
 
     import spark.implicits._
-    import za.co.absa.abris.examples.utils.ExamplesUtils._
 
     implicit val encoder: Encoder[Row] = getEncoder(properties)
 
@@ -62,9 +63,7 @@ object KafkaAvroWriterWithKey {
 
       val dataframe = spark.sparkContext.parallelize(rows, properties.getProperty(PARAM_NUM_PARTITIONS).toInt).toDF()
 
-      dataframe.printSchema()
-
-      toAvro(dataframe, properties)
+      toAvro(dataframe, properties) // check the method content to understand how the library is invoked
         .write
         .format("kafka")
         .addOptions(properties) // 1. this method will add the properties starting with "option."; 2. security options can be set in the properties file
@@ -73,14 +72,25 @@ object KafkaAvroWriterWithKey {
   }
 
   private def toAvro(dataframe: Dataset[Row], properties: Properties) = {
+
     import za.co.absa.abris.avro.AvroSerDeWithKeyColumn._
 
+    val sc = Map(
+      SchemaManager.PARAM_SCHEMA_REGISTRY_URL -> properties.getProperty(SchemaManager.PARAM_SCHEMA_REGISTRY_URL),
+      SchemaManager.PARAM_KEY_SCHEMA_NAMING_STRATEGY -> properties.getProperty(SchemaManager.PARAM_KEY_SCHEMA_NAMING_STRATEGY),
+      SchemaManager.PARAM_VALUE_SCHEMA_NAMING_STRATEGY -> properties.getProperty(SchemaManager.PARAM_VALUE_SCHEMA_NAMING_STRATEGY)
+    )
+
+    val topic = properties.getProperty(PARAM_TOPIC)
+
+    // providing access to Schema Registry is mandatory
     if (properties.getProperty(PARAM_INFER_SCHEMA).trim().toBoolean) {
-      val name = properties.getProperty(PARAM_AVRO_RECORD_NAME)
-      val namespace = properties.getProperty(PARAM_AVRO_RECORD_NAMESPACE)
-      dataframe.toAvro(name, namespace, name, namespace)
-    } else {
-      dataframe.toAvro(properties.getProperty(PARAM_KEY_AVRO_SCHEMA), properties.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA))
+      val schemaName = properties.getProperty(PARAM_AVRO_RECORD_NAME)
+      val schemaNamespace = properties.getProperty(PARAM_AVRO_RECORD_NAMESPACE)
+      dataframe.toConfluentAvro(topic, schemaName, schemaNamespace, schemaName, schemaNamespace)(sc)
+    }
+    else {
+      dataframe.toConfluentAvro(topic, properties.getProperty(PARAM_KEY_AVRO_SCHEMA), properties.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA))(sc)
     }
   }
 
@@ -88,11 +98,10 @@ object KafkaAvroWriterWithKey {
     var count = 0
     ComplexRecordsGenerator
       .generateUnparsedRows(howMany)
-        .map(row => {
-          count = count + 1
-          Row(Row(count, s"whatever string $count"),row)
-        })
-
+      .map(row => {
+        count = count + 1
+        Row(Row(count, s"whatever string $count"),row)
+      })
   }
 
   private def getEncoder(properties: Properties): Encoder[Row] = {
