+++ src/main/scala/za/co/absa/abris/examples/deprecated/using_keys/KafkaAvroReaderWithKey.scala
@@ -14,19 +14,22 @@
  * limitations under the License.
  */
 
-package za.co.absa.abris.examples.sql
+package za.co.absa.abris.examples.deprecated.using_keys
 
 import java.util.Properties
 
-import org.apache.spark.sql.DataFrame
-import org.apache.spark.sql.functions.col
+import org.apache.spark.sql.streaming.DataStreamReader
+import org.apache.spark.sql.{Dataset, Row}
+import za.co.absa.abris.avro.schemas.policy.SchemaRetentionPolicies.RETAIN_ORIGINAL_SCHEMA
 import za.co.absa.abris.examples.utils.ExamplesUtils._
 
-object NewKafkaAvroReader {
+object KafkaAvroReaderWithKey {
 
   private val PARAM_JOB_NAME = "job.name"
   private val PARAM_JOB_MASTER = "job.master"
+  private val PARAM_KEY_AVRO_SCHEMA = "key.avro.schema"
   private val PARAM_PAYLOAD_AVRO_SCHEMA = "payload.avro.schema"
+  private val PARAM_TASK_FILTER = "task.filter"
   private val PARAM_LOG_LEVEL = "log.level"
   private val PARAM_OPTION_SUBSCRIBE = "option.subscribe"
 
@@ -34,7 +37,6 @@ object NewKafkaAvroReader {
 
   def main(args: Array[String]): Unit = {
 
-    // check if properties file is present, exists if not
     // there is an example file at /src/test/resources/AvroReadingExample.properties
     checkArgs(args)
 
@@ -45,33 +47,30 @@ object NewKafkaAvroReader {
     val stream = spark
       .readStream
       .format("kafka")
-      .addOptions(properties) // 1. this method will add the properties starting with "option."
-                              // 2. security options can be set in the properties file
+      .addOptions(properties) // 1. this method will add the properties starting with "option."; 2. security options can be set in the properties file
 
-    val deserialized = configureExample(stream.load(), properties)
+    val deserialized = configureExample(stream, properties)
 
     // YOUR OPERATIONS CAN GO HERE
 
     deserialized.printSchema()
+
     deserialized
       .writeStream
       .format("console")
-      .option("truncate", "false")
       .start()
       .awaitTermination()
   }
 
-  private def configureExample(dataFrame: DataFrame, props: Properties): DataFrame = {
-    import za.co.absa.abris.avro.functions.from_avro
+  private def configureExample(stream: DataStreamReader,props: Properties): Dataset[Row] = {
+
+    import za.co.absa.abris.avro.AvroSerDeWithKeyColumn._
 
     if (props.getProperty(PARAM_EXAMPLE_SHOULD_USE_SCHEMA_REGISTRY).toBoolean) {
-      val schemaRegistryConfig = props.getSchemaRegistryConfigurations(PARAM_OPTION_SUBSCRIBE)
-      dataFrame.select(from_avro(col("value"), schemaRegistryConfig) as 'data).select("data.*")
+      stream.fromAvro(props.getSchemaRegistryConfigurations(PARAM_OPTION_SUBSCRIBE))(RETAIN_ORIGINAL_SCHEMA)
     }
     else {
-      val source = scala.io.Source.fromFile(props.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA))
-      val schemaString = try source.mkString finally source.close()
-      dataFrame.select(from_avro(col("value"), schemaString) as 'data).select("data.*")
+      stream.fromAvro(props.getProperty(PARAM_KEY_AVRO_SCHEMA), props.getProperty(PARAM_PAYLOAD_AVRO_SCHEMA))(RETAIN_ORIGINAL_SCHEMA)
     }
   }
 }
